Migrating tests, changes to integration tests
We modularized the flask app, so we changed over the tests to point to the new module. And other changes to setup the app to run correctly when being tested.

Then, since we implemented an optional command line argument for where to save the database, the integration tests were changed to save the database in a temp directory. This avoids having the database have a consistent state between tests.