Made changes to make a package
Moved code to src, and added relative imports.

Then modified the testing to point to the new directory.

Then modified main.py to act as a main entrypoint for running the script outside of a package.

Then setup setup.py and MANIFEST to allow for package building.

Then test deployed package on Ubuntu VM, confirmed testing, and verified working.