modify code to use new validation and standardization class
Instead of using a single validation method in Database, I moved the validation and expanded it to a new class.

Then, input should have additional information appended to it (e.g., timestamp), and so that functionality was added as well.

Consequently these changes and the new input requirements necessitated a change in the testing; and so the testing was modified as well.